% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidykids_dictionary.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary for the tidykids dataset}
\usage{
data_dictionary()
}
\value{
A tibble with the variable descriptions
}
\description{
Data dictionary for the tidykids dataset
}
