% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidykids.R
\docType{data}
\name{tidykids}
\alias{tidykids}
\title{State-by-State Spending on Kids Dataset}
\format{
A data frame with 70380 rows and 4 variables:
\describe{
\item{state}{United States state (and the District of Columbia), a string}
\item{variable}{Variable, see tidykids_data_dictionary for more information, a string}
\item{year}{year, an integer}
\item{raw}{the value of the variable; a numeric value}
\item{inf_adj}{the value of the variable, adjusted for inflation, a numeric value}
\item{inf_adj_perchild}{the value of the variable adjusted for inflation, per child; a numeric value}
}
}
\source{
\url{https://datacatalog.urban.org/dataset/state-state-spending-kids-dataset}
}
\usage{
tidykids
}
\description{
A dataset containing state-by-year data on spending on kids in the United States, 1997-2016. The data was made available by and is attributable to Julia Isaacs, Eleanor Lauderback, and Erica Greenberg at the Urban Institute under the under the ODC Attribution License (https://opendatacommons.org/licenses/by/1-0/).
}
\keyword{datasets}
